requirements.txt
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
executing==2.1.0
fonttools==4.55.3
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
z3-solver==4.13.4.0
